+++ content/signRulesDefaultsView.mjs.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020-2022 Philippe Lieser
+ * Copyright (c) 2020-2022;2024 Philippe Lieser
  *
  * This software is licensed under the terms of the MIT License.
  *
@@ -10,20 +10,64 @@
 // @ts-check
 /* eslint-env webextensions */
 
-import DataTable from "./table.mjs.js";
 import ExtensionUtils from "../modules/extensionUtils.mjs.js";
 import SignRulesProxy from "../modules/dkim/signRulesProxy.mjs.js";
+import { TabulatorFull as Tabulator } from "../thirdparty/tabulator-tables/dist/js/tabulator_esm.js";
 import { getElementById } from "./domUtils.mjs.js";
 
 document.addEventListener("DOMContentLoaded", async () => {
-	const tableElement = getElementById("rulesTable");
-	if (!(tableElement instanceof HTMLTableElement)) {
-		throw new Error("element rulesTable is not a HTMLTableElement");
-	}
+	// Initialize table
 
 	const data = await SignRulesProxy.getDefaultRules();
-	const table = new DataTable(tableElement);
-	table.showData(data);
+
+	const table = new Tabulator("#rulesTable", {
+		height: "100%",
+		data,
+		// Note: The virtual renderer would be nicer, but there are multiple scrolling issues.
+		renderVertical: "basic",
+		layout: "fitColumns",
+		columns: [
+			{
+				title: browser.i18n.getMessage("treeviewSigners.treecol.domain"),
+				field: "domain",
+				formatter: "textarea",
+			},
+			{
+				title: browser.i18n.getMessage("treeviewSigners.treecol.addr"),
+				field: "addr",
+				formatter: "textarea",
+			},
+			{
+				title: browser.i18n.getMessage("treeviewSigners.treecol.sdid"),
+				field: "sdid",
+				formatter: "textarea",
+			},
+			{
+				title: browser.i18n.getMessage("treeviewSigners.treecol.ruletype"),
+				field: "type",
+				maxWidth: 130,
+			},
+			{
+				title: browser.i18n.getMessage("treeviewSigners.treecol.priority"),
+				field: "priority",
+				maxWidth: 130,
+			},
+		],
+		initialSort: [
+			{ column: "domain", dir: "asc" },
+		],
+	});
+
+	// Workaround for https://github.com/olifolkerd/tabulator/issues/4277
+	table.eventBus?.subscribe("table-redraw", (/** @type {boolean} */ force) => {
+		if (!force) {
+			for (const row of table.getRows()) {
+				row.normalizeHeight();
+			}
+		}
+	});
+
+	// Initialize buttons
 
 	const buttonHelp = getElementById("buttonHelp");
 	buttonHelp.addEventListener("click", () => {