+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Enhancements
+
+- Improved table views for sign rules and DKIM keys (#248, #305).
+  E.g. it is now possible to delete multiple entries at once.
+
 ## 5.4.0 (2023-11-16)
 
 ### Enhancements